+++ src/test/java/roomescape/repository/ReservationThemeDbRepositoryTest.java
@@ -1,45 +1,25 @@
 package roomescape.repository;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.SoftAssertions.*;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.context.ActiveProfiles;
-import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTheme;
 import roomescape.repository.dao.ReservationThemeH2Dao;
+import roomescape.support.JdbcTestSupport;
 
-@JdbcTest
-@ActiveProfiles("test")
 @Import({ReservationThemeDbRepository.class, ReservationThemeH2Dao.class})
-class ReservationThemeDbRepositoryTest {
+class ReservationThemeDbRepositoryTest extends JdbcTestSupport {
 
     @Autowired
     private ReservationThemeDbRepository reservationThemeDbRepository;
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
-
-    @BeforeEach
-    void setUp() {
-        jdbcTemplate.update("""
-                SET REFERENTIAL_INTEGRITY FALSE;
-                TRUNCATE TABLE reservation;
-                ALTER TABLE reservation ALTER COLUMN id RESTART WITH 1;
-                TRUNCATE TABLE reservation_time;
-                ALTER TABLE reservation_time ALTER COLUMN id RESTART WITH 1;
-                TRUNCATE TABLE theme;
-                ALTER TABLE theme ALTER COLUMN id RESTART WITH 1;
-                SET REFERENTIAL_INTEGRITY TRUE;
-                """);
-    }
 
     @DisplayName("존재하지 않는 id를 조회하면 예외를 발생시킨다")
     @Test
